# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "XternityGames Web3 API Documentation",
    "version": "1.0.0",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://dev-api.xternity.games/v1"
    }
  ],
  "paths": {
...
    "/me": {
      "get": {
        "summary": "Get player info",
        "tags": [
          "Player"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "results": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Player"
                          }
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": null
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/me/assets": {
      "get": {
        "summary": "Get player minted nfts inventory",
        "tags": [
          "Player"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "orderBy",
            "schema": {
              "type": "string",
              "default": "asc"
            },
            "description": "sort by query in the form of field:desc/asc (ex. name:asc)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "number",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "results": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Assets"
                          }
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "number"
                            },
                            "limit": {
                              "type": "number"
                            },
                            "totalPages": {
                              "type": "number"
                            },
                            "totalResults": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Player": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "externalPlayerId": {
            "type": "string"
          },
          "gameId": {
            "type": "string"
          },
          "isEmailVerified": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "wallets": {
            "type": "object",
            "properties": {
              "polygon": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "string"
                  },
                  "isExternal": {
                    "type": "boolean"
                  }
                }
              }
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "init",
              "active"
            ]
          }
        }
      },
...
      "Assets": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "network": {
            "type": "string"
          },
          "collection": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "icon_url": {
                "type": "string"
              }
            }
          },
          "created_at": {
            "type": "string"
          },
          "synced_at": {
            "type": "string"
          },
          "image_url": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "image": {
                "type": "string"
              },
              "health": {
                "type": "number"
              }
            }
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "token_hash": {
            "type": "string"
          },
          "token_address": {
            "type": "string"
          },
          "token_uri": {
            "type": "string"
          },
          "block_number_minted": {
            "type": "string"
          },
          "block_number": {
            "type": "string"
          },
          "token_id": {
            "type": "string"
          },
          "is_valid": {
            "type": "number"
          },
          "updated_at": {
            "type": "string"
          },
          "uri": {
            "type": "string"
          },
          "user": {
            "type": "string"
          },
          "owner_of": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "contract_type": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "syncing": {
            "type": "number"
          },
          "frozen": {
            "type": "number"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "ERROR"
          },
          "data": {
            "type": "object"
          },
          "meta": {
            "type": "object"
          },
          "error": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": "ERROR",
              "data": {},
              "meta": {},
              "error": [
                "Please authenticate"
              ]
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": "ERROR",
              "data": {},
              "meta": {},
              "error": [
                "Forbidden"
              ]
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}