# src/routes/v1/player-verse/player.route.ts
import { Router } from 'express';
import { auth } from '@/middlewares/auth';
import { useGame } from '@/middlewares/game';
import { authApiKey } from '@/middlewares/auth-api-key';
import { extractSortAndPagination } from '@/middlewares/sortAndPagination';
import { PlayerController } from '@/controllers/player-verse/player.controller';

const router = Router();

router.route('/me').get(auth(), authApiKey, PlayerController.getPlayerInfo);

router.route('/me/assets').get(auth(), useGame, extractSortAndPagination, PlayerController.getPlayerInventory);

export const PlayerRoute = {
  router,
};

/**
 * @swagger
 * tags:
 *   name: Player
 *   description: Players API
 */

/**
 * @swagger
 * /me:
 *   get:
 *     summary: Get player info
 *     tags: [Player]
 *     security:
 *       - bearerAuth: []
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               token:
 *                 type: string
 *     responses:
 *       "200":
 *         description: OK
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: string
 *                   example: "OK"
 *                 data:
 *                   type: object
 *                   properties:
 *                     results:
 *                       type: array
 *                       items:
 *                         $ref: '#/components/schemas/Player'
 *                 meta:
 *                   type: object
 *                   properties:
 *                 error:
 *                   type: string
 *       "401":
 *         $ref: '#/components/responses/Unauthorized'
 *       "403":
 *         $ref: '#/components/responses/Forbidden'
 */

/**
 * @swagger
 * /me/assets:
 *   get:
 *     summary: Get player minted nfts inventory
 *     tags: [Player]
 *     security:
 *       - bearerAuth: []
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               token:
 *                 type: string
 *     parameters:
 *       - in: query
 *         name: orderBy
 *         schema:
 *           type: string
 *           default: "asc"
 *         description: sort by query in the form of field:desc/asc (ex. name:asc)
 *       - in: query
 *         name: limit
 *         schema:
 *           type: number
 *           default: 10
 *       - in: query
 *         name: page
 *         schema:
 *           type: integer
 *           minimum: 1
 *           default: 1
 *         description: Page number
 *     responses:
 *       "200":
 *         description: OK
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: string
 *                   example: "OK"
 *                 data:
 *                   type: object
 *                   properties:
 *                     results:
 *                       type: array
 *                       items:
 *                         $ref: '#/components/schemas/Assets'
 *                 meta:
 *                   type: object
 *                   properties:
 *                     pagination:
 *                       type: object
 *                       properties:
 *                         page:
 *                           type: number
 *                         limit:
 *                           type: number
 *                         totalPages:
 *                           type: number
 *                         totalResults:
 *                           type: number
 *                 error:
 *                   type: string
 *       "401":
 *         $ref: '#/components/responses/Unauthorized'
 *       "403":
 *         $ref: '#/components/responses/Forbidden'
 */